requirements.txt
streamlit==1.18.0
python-telegram-bot==21.7
nest-asyncio==1.5.6
altair==4.2.0
vega-datasets==0.9.0
flask==3.0.3
